Refactor RNTupleFieldPtr to thin wrapper
RNTupleFieldPtrs will have to be used by more than just
FlatTables, so refactor it to accept a simple value and
leave filling it up to the caller.
